[add]添加收获功能，修复报错

+++ Assets/Scripts/LittleWorld/Nature/PlantInfo.cs
@@ -16,7 +16,7 @@ namespace LittleWorld.Item
         public int maxHealth;
         public int seedItem;
         public float nutrition;
-        public int growingTime;
+        public float growingTime;
         public List<Sprite> itemSprites;
     }
 }
