+++ Assets/Scripts/Xml/XmlUtility.cs
@@ -35,7 +35,7 @@ namespace Xml
                     plant.maxHealth = int.Parse(item.SelectSingleNode("maxHealth").InnerText);
                     plant.seedItem = int.Parse(item.SelectSingleNode("seedItem").InnerText);
                     plant.nutrition = float.Parse(item.SelectSingleNode("nutrition").InnerText);
-                    plant.growingTime = int.Parse(item.SelectSingleNode("growingTime").InnerText);
+                    plant.growingTime = float.Parse(item.SelectSingleNode("growingTime").InnerText);
                     plant.itemSprites = CreateItemSpritesList(item, 6);
                     ObjectConfig.plantInfoDic.Add(plant.itemCode, plant);
                 }
